Resolve PR
