+++ src/entities/attachment/api/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from "./use-fetch-by-id-v1";
+export * from "./use-search-v1";
