+++ src/entities/attachment/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./api-service";
+export * from "./hooks";
