+++ src/entities/attachment/libs/index.ts
@@ -1,0 +1,3 @@
+export * from "./helpers";
+export * from "./hooks";
+export * from "./types";
