+++ src/entities/twin/api/api-service.ts
@@ -1,8 +1,6 @@
 import { PaginationState } from "@tanstack/table-core";
 
 import {
-  AttachmentFilters,
-  AttachmentRqQuery,
   TwinCreateRq,
   TwinFilters,
   TwinSimpleFilters,
@@ -179,54 +177,6 @@ export function createTwinApi(settings: ApiSettings) {
     );
   }
 
-  function searchAttachment({
-    pagination,
-    filters,
-  }: {
-    pagination: PaginationState;
-    filters: AttachmentFilters;
-  }) {
-    return settings.client.POST("/private/attachment/search/v1", {
-      params: {
-        header: getApiDomainHeaders(settings),
-        query: {
-          lazyRelation: false,
-          showAttachmentMode: "DETAILED",
-          showAttachment2TwinMode: "DETAILED",
-          showAttachment2TransitionMode: "DETAILED",
-          showAttachment2UserMode: "DETAILED",
-          showAttachment2PermissionMode: "DETAILED",
-          showAttachment2CommentModeMode: "DETAILED",
-          showAttachmentCollectionMode: "ALL",
-          showTwinClass2TwinClassFieldMode: "DETAILED",
-          showTwin2TwinClassMode: "DETAILED",
-          offset: pagination.pageIndex * pagination.pageSize,
-          limit: pagination.pageSize,
-          sortAsc: false,
-        },
-      },
-      body: {
-        ...filters,
-      },
-    });
-  }
-
-  function getAttachment({
-    attachmentId,
-    query,
-  }: {
-    attachmentId: string;
-    query: AttachmentRqQuery;
-  }) {
-    return settings.client.GET("/private/attachment/{attachmentId}/v1", {
-      params: {
-        header: getApiDomainHeaders(settings),
-        path: { attachmentId },
-        query,
-      },
-    });
-  }
-
   return {
     search,
     // getById,
@@ -237,8 +187,6 @@ export function createTwinApi(settings: ApiSettings) {
     getLinks,
     upsertField,
     getValidTwinsForLink,
-    searchAttachment,
-    getAttachment,
   };
 }
