+++ src/screens/twin/views/index.ts
@@ -1,3 +1,4 @@
+export * from "./twin-attachments";
 export * from "./twin-fields";
 export * from "./twin-general";
 export * from "./twin-links";
