+++ src/screens/twin/views/twin-attachments.tsx
@@ -1,0 +1,3 @@
+// TODO: implement TwinAttachments
+// via return <AttachmentsTable /> with specific props
+export const TwinAttachments = () => <span>WIP</span>;
